requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-rest-passwordreset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
idna==2.10
inflection==0.3.1
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
more-itertools==8.5.0
openapi-codec==1.3.2
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
